Add resident home, collection events and live fan-out (T034-T041)
Collection events are the heart of this: when an active trip's accepted
position comes within 75 m of a house, the system records that the house was
served. Nobody presses anything, and the evidence is the GPS trail. That turns
"last collected" from a claim into a fact, and it is what a missed-pickup
dispute is settled against. Verified with the simulator — all three seeded
households recorded themselves as the auto drove past.

Fixes a bug that only showed up by reading the response: pg returns `numeric`
as a string, so `extract(epoch ...)` fed `new Date()` an Invalid Date. The home
screen claimed a collection had happened while reporting no timestamp for it.
All three call sites now cast to float8; a regression test asserts the returned
value is a valid Date.

WebSocket scoping is enforced by JWT claim rather than broker ACL: a resident
may subscribe only to the route their household sits on. That is a one-line
comparison an auditor can check, where an ACL rule set is not. Sockets outlive
the 15-minute token, so they are cycled with a reauth frame at 60 minutes.

Ingest gained a listener hook so collection recording and live fan-out attach
without ingest depending on the modules that consume it. Listener failures are
logged, never propagated: a dropped notification is recoverable, a lost ping is
not.

64 API tests, 12 Flutter tests, analyze clean.

+++ apps/api/src/modules/geo/households.service.ts
@@ -26,7 +26,7 @@ export class HouseholdsService {
         "ward_id",
         // Epoch millis, so the aging comparison does not depend on how the
         // driver hands back timestamps.
-        sql<number>`extract(epoch from created_at) * 1000`.as("created_ms"),
+        sql<number>`(extract(epoch from created_at) * 1000)::float8`.as("created_ms"),
       ])
       .where("mapping_status", "=", "pending_review")
       .$if(Boolean(wardId), (qb) => qb.where("ward_id", "=", wardId))
